3-hard/requirements.txt
sockets
requests
pytube3